DockerManager/requirements.txt
Django==1.11.7
pycrypto==2.6.1
paramiko==2.4.1
pymysql==0.9.0
django-extensions==2.0.7
django-werkzeug-debugger-runserver==0.3.1
pyOpenSSL==18.0.0
